91_deploy/requirements.txt
-i https://pypi.python.org/simple
alembic==1.13.1; python_version >= '3.8'
aniso8601==9.0.1
appnope==0.1.4; sys_platform == 'darwin'
asttokens==2.4.1
backcall==0.2.0
bcrypt==4.1.2; python_version >= '3.7'
blinker==1.7.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
decorator==5.1.1; python_version > '3.6' and python_version < '3.11'
executing==2.0.1; python_version >= '3.5'
faker==23.3.0
flask==3.0.2
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-migrate==4.0.5
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
gunicorn==21.2.0
honcho==1.1.0
importlib-metadata==7.0.1; python_version < '3.10'
importlib-resources==6.1.2; python_version < '3.9'
ipdb==0.13.13
ipython==8.12.3; python_version > '3.6' and python_version < '3.11'
itsdangerous==2.1.2; python_version >= '3.7'
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.3; python_version >= '3.7'
mako==1.3.2; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
matplotlib-inline==0.1.6; python_version >= '3.5'
packaging==23.2; python_version >= '3.7'
parso==0.8.3; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.43; python_version >= '3.7'
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.17.2; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlalchemy==2.0.27
sqlalchemy-serializer==1.4.1
stack-data==0.6.3
tomli==2.0.1; python_version > '3.6' and python_version < '3.11'
traitlets==5.14.1; python_version >= '3.8'
typing-extensions==4.10.0; python_version <= '3.8'
wcwidth==0.2.13
werkzeug==3.0.1; python_version >= '3.8'
zipp==3.17.0; python_version >= '3.8'
